Use binary resources for the 'nativeruntime' project
Previously nativeruntime tests would not execute in SDK > 28 in Gradle
as they were not configured to run in binary resources.

To use binary resources for a project, a test_config.properties file has
to be added to the test resources, and a resources.ap_ file has to be
generated using the scripts/build-resources.sh script.