Introduce a separate system property to enable real scrolling code
By default, when using legacy graphics, Robolectric has a custom implementation
of View scrolling logic that is broken in some key areas. For instance, in real
Android, a ScrollView with no children is not scrollable using
`smoothScrollTo`, but it is in Robolectric.

Introduce a system property that can control whether to use real or fake
scrolling code. The eventual goal is to always use real scrolling
implementation by default to make tests more realistic. However, this may take
time and investigation as many tests rely on the broken behavior.

PiperOrigin-RevId: 503066727