Filter out window configuration changes when recreating an activity
In the Android window manager these change flags are filtered out and do not trigger an activity recreate by themselves. This is causing a bug in Robolectric where activity's are being recreated when they should be receiving a configuration change instead e.g. when they handle rotation but the window is rotated.

PiperOrigin-RevId: 504875595