Add a system property to only use real spinner selection code
The AbsSpinner.setSelection(int) shadow method is broken in multiple
ways. It causes the item selection callback to be invoked in many
situations where it shouldn't be invoked, leading to low-fidelity behavior
and infinite loops. However, there are a lot of tests that depend
on this broken behavior.

Add a system property to control whether only real spinner selection code
is used. It is false for now and at some point will be switched to true.

PiperOrigin-RevId: 509884692