Fix an edge cases in Jacoco detection logic for constructors
Previously, there was a check for Jacoco-instrumented constructors that only
looked at the first instruction. However, it's possible that the first
instruction is a LabelNode. Add a check for the second instruction as well.

There still may be some edge cases here, but it's better to wait to see those
before trying to come up with a generic solution.

Just for some additional context, Robolectric instrumentation is skipped
on Jacoco-instrumented constructors, as the breaking apart of constructors
that Robolectric does invalidates Jacoco load instructions.

PiperOrigin-RevId: 515189416