Add a gradle task to publish the nativeruntime compat distribution to MavenCentral
This jar will contain all assets for the native runtime, including shared
object files for Linux and Mac, ICU data, and fonts. It contains native
libraries for Android SQLite and Android graphics.

The reason it is named 'compat' is that this version is compatible with
multiple SDK levels, currently O through T. However, starting in Android U or
Android V, we would like to move to a model where each Android SDK Jar will
also contain its own native libraries for SQLite, Graphics, and potentially
other subsystems like Resources, Binder, Javacore, etc..

This task is gated on an environment variable so that it is not published
by default when `./gradlew publish` is called.

PiperOrigin-RevId: 516429231