Fix ShadowLooper.idlefor deadlocks.
This commit fixes this issue by introducing a temporary set of static listeners
that are not reset between tests. This is necessary since Loopers and their
MessageQueues are reused between test, and won't get registered as listeners for
each new test.

A future change will attempt to reduce this to one set of listeners,
which will involve cleaning up a bunch of issues in ShadowChoreographer and friends.

PiperOrigin-RevId: 522125697