Remove precondition check in ShadowActivityThread resetter
Previously, the ShadowActivityThread resetter would crash in a precondition
check if the ActivityThread has not been set.

Generally the ActivityThread should be getting set, except if an exception
occurs during test initialization. if an exception does occur early in
initialization, it is not very useful to throw an exception in the resetter, as
it causes the original exception message to be replaced with the Exception.

Instead, just skip the resetter logic if the ActivityThread is detected to be
null.

Updates #8160

PiperOrigin-RevId: 527803222