Add support for invoking constructors with reflector
Previously, it was not possible to invoke constructors using a reflector
interface. Constructors that were either hidden or inaccessible (e.g. in a
different API level) had to be invoked using ReflectionHelpers, which is
tedious and slow. Add support for invoking constructors in reflector with the
`@Constructor` annotation.

Also, migrate the two highest-usage ReflectionHelpers constructor usages to
reflector.

Using reflector to invoke constructors is 10-100x faster than using
ReflectionHelpers, depending on whether the reflector object is cached.

PiperOrigin-RevId: 529310268