Allow creation of AudioTrack instances for offload
With this change, AudioTrack instances can be created for offload mode
as long as the platform supports offload playback is supported for the
specific encoding.

Offload was introduced in API 29 and the check for offload support is
done in AudioTrack.Builder.build(), however the check is implemented
differently through the SDK versions:
- AudioSystem.isOffloadSupported() for APIs 29-31
- AudioSystem.getOffloadSupport() for API 32
- AudioSystem.getDirectPlaybackSupport() for API 33+

The commit adds helper methods in ShadowAudioSystem so that
each AudioSystem API is controlled separately. That's because these APIs
have a  different native source of truth and it's possible that the APIs
return different values for the same input.

In the same spirit, this commit adds one more helper method in
ShadowAudioTrack to control separately the creation of AudioTrack
instances for passthrough playback which were previously coupled with
AudioTrack.isDirectPlaybackSupported(). However the native code does
check with AudioTrack.isDirectPlaybackSupported() when creating an
AudioTrack instance.

PiperOrigin-RevId: 530946885