Ignore 'PiperOrigin-RevId' line when validating commit message body
Nearly 45% of commit messages that originate from Google via Copybara are
one-liners no commit message body. A GitHub action was added recently to
address this, but it considered the 'PiperOrigin-RevId' tag as a line in the
commit message body, so all commits via Copybara were considered valid.

A descriptive commit message is important in many ways, especially when
debugging problems or refactoring code.

PiperOrigin-RevId: 547389888