Add 'startObservingDevicePresence' and the systemApi version of 'associate' to ShadowCompanionDeviceManager.
The API `startObservingDevicePresence` does not return a result, but it does throw an exception if an association with the device is not made. By adding a `startObservingDevicePresence` shadow implementation, tests can assert the behavior of the API when an association has not be made. Regarding the system API version of 'associate', for system applications and services that use this API, an implementation is added to be able to test against this version of associate.

PiperOrigin-RevId: 557879309