Fix a memory leak in ShadowVMRuntime
Previously, ShadowVMRuntime, which is effectively a singleton, used
NativeAllocationRegistry but contained no logic to free references. This meant
that it retained a reference for any object O for which VMRuntime.addressOf(O)
was called. This happened with all TypedArray objects via TypedArray.obtain,
which are heavily used in the Android resource system.

Switch to weak references which allows these objects to be collected.

For #8405

PiperOrigin-RevId: 558252950