Fix ShadowPixelCopy cropping behavior.
ShadowPixelCopy.request(Surface) simulates Android behavior by
grabbing the surface's corresponding decorView, and capturing
an image of it. But this approach doesn't accurately account
of cases where the decorView is offset within the surface,
such as when capturing images of views inside a Dialog.

This commit adjusts the Rect cropping provided to ShadowPixelCopy
to adjust it by the decorView's position in its surface.

PiperOrigin-RevId: 558941353