Add accessor methods to return the inputs of startObservingDevicePresence and associate of ShadowCompanionDeviceManager
This change makes it simpler to test interactions with CompanionDeviceManager APIs
startObservingDevicePresence and the sytemApi version of 'associate' by adding
accessor methods that return the last device address passed to these APIs. Both
of theses APIs do not return a result, so by adding these accessor methods, a
test can verify that these methods were executed or not executed.

PiperOrigin-RevId: 559231109