Avoid unnecessary creation of sandboxes when custom shadows are used
In Robolectric, a new sandbox was created whenever there was a change in the
instrumented class or package lists in InstrumentationConfiguration.

However, there were no redundancy checks. If one test specified
'@Config(intrumentedPackages={"android."})', a new sandbox would be created,
even though 'android.' is already configured to be instrumented.  Similarly, a
new sandbox was created whenever a test specified custom shadows using
`@Config(shadows=...)`, even though InvokeDyamic SwitchPoints can be used to
rewire Android classes to use custom shadows.

Update the InstrumentationConfiguration builder to optimize the instrumented
package and instrumented classes to support this case.

This change may impact tests that have custom shadows for static variables that
leak across tests. However, this is a very rare occurrence, and it is easily
mitigated.

For #8432
For #7359

PiperOrigin-RevId: 560522727