Remove redundant caching of ApkAssets objects in ShadowArscAssetManager{9,10}
There is already a static cache of ApkAssets objects in ShadowArscApkAssets9,
so an additional layer of caching is redundant. It was verified that this does
not result in system resources being loaded more times.

AssetManager.createSystemAssetsInZygoteLocked ends up calling
ApkAssets.loadFromPath, which is where the other ApkAssets cache
is located (See ShadowArscApkAssets9.loadFromPath methods).

PiperOrigin-RevId: 561227074