Support setting accessibility timeouts.
This allows configuring the return value of `getRecommendedTimeoutMillis(int, int)`.

This change has nothing to do with faking time. There is a feature of the AccessibilityManager / the [Android Accessibility Settings](https://support.google.com/accessibility/android/answer/9426889) for the user to select their desired timeouts for UIs. This is used by the developer to create inclusive and accessible UIs that provide the user their preferred amount of time to make a decision about ephemeral UI content. Developers should be using (and testing) their usage of `getRecommendedTimeoutMillis` when creating these affordances.

This change will allow developers to change the user setting to be configured for testing purposes. Currently, the A11yManager always returns 0 in Robolectric tests.

PiperOrigin-RevId: 561518834