Use a global cache for ApkAssets
Previously there was a per-sandbox cache for ApkAssets objects that was stored
redundantly in both ShadowArscApkAssets9 and ShadowArscAssetManager{9,10}.
However, with the proliferation of sandbox parameters (looper modes, graphics
modes, instrumentation configuration), and the increase in size of system
resource tables, system resource blobs began to dominate the memory consumption
of test processes. This problem was especially salient in large multi-sdk test
processes common in Gradle environments.

Switch to using a global cache for system resources. This will reduce overall
memory consumption for tests that initialize a lot of sandboxes. It also makes
it possible for ApkAssets objects to be collected (though the underlying
CppApkAssets will not be collected). If necessary, there could be logic to
clear this cache when sandboxes of a certain SDK level are evicted from the
SandboxManager cache.

For #7359
For #8432

PiperOrigin-RevId: 561711131