Remove obsolete TestExecutionContext class
The TestExecutionContext was a way to store certain data required to run
Robolectric tests as static variables. It was added at a time when
RoboFrameworkMethod objects were required to support being serialized through
XStream. Although XStream was used in the past to serialize parameters for
ParameterizedRobolectricTestRunner, it is no longer used any more. Remove this
class and store the data as members on RoboFrameworkMethod.

Also remove an obsolete test. There will be forthcoming commits that add a
larger number of GC-related tests for Robolectric sandboxes.

PiperOrigin-RevId: 562203698