Adapt to new signatures for measuring text, telephony manager and asset manager.
The in development android platform
has modified signatures for several methods relating to measuring text, telephony manager and asset manager, which were being shadowed. This causes Robolectric's
annotation validator to fail when inspecting the new android all jar.

This commit is a quick workaround to modify the shadows
to add variants for the new signatures, to placate the annotation validator.

A future change should consider
adding full support for the new `AssetManager::setConfiguration`, `TelephponyManager::setPhoneType` and `MeasuredText::nAddStyleRun` functionality.

PiperOrigin-RevId: 566399788