Only advance the looper manually if it is paused.
If the looper is not paused it will advance itself. The default behavior is that in `INSTRUMENTATION_TEST` looper mode the main looper is not paused. The previous logic, which called `runMainLooperToNextTask`, would create a race the looper where if the looper hadn't finished running the posted runnable then the subsequent message on the looper would be run before the method returned.

PiperOrigin-RevId: 568247277