Add an implementation of `dup()` to ShadowParcelFileDescriptor.
This is to fix the following situation, which currently fails:
```
File testFile = new File(context.getFilesDir(), "test");
ParcelFileDescriptor fd =
    ParcelFileDescriptor.open(testFile, ParcelFileDescriptor.MODE_READ_WRITE);

ParcelFileDescriptor dupFd = fd.dup();
FileDescriptor file = dupFd.getFileDescriptor(); // Fails with NPE.
```

This happens because dup is not implemented in the shadow, so dupFd's shadow
does not have the `file` field set. When invoking `dupFd.getFileDescriptor()`,
it tries to invoke `getFile().getFd()` which fails with a NullPointerException.

By implementing a shadow for `dup`, it can simply call the constructor which
sets this `file` field as expected.

PiperOrigin-RevId: 573024247