Add ShadowPausedLooper.postSync
This method allows the caller to post a runnable to the looper and then block on returning until after the runnable has completed, and will rethrow any uncaught exception that was thrown on the main thread.

Previously in instrumentation test looper mode calling Instrumentation.runOnMainSync would simply block if the main thread had uncaught exceptions.

PiperOrigin-RevId: 573041512