Update GitHub CI workflows for Copybara PRs on the 'google' branch
When a Robolectric change is made by a Googler, all Robolectric tests are
already run on Google's infrastructure. It is redundant to re-run tests on
GitHub actions.

With this change, when a PR is made by the Copybara service:
1) The PR will be built to ensure that there are no compile errors.
2) All integration tests will be run on the latest SDK version.
3) GitHub CI tests will be run when the commit is pushed to the 'google' branch.

This will reduce the amount of contention for GitHub CI and will significantly
decrease the amount of time required for Copybara PRs to be merged.

Also, update AndroidProjectConfigPlugin.groovy to avoid running tests on
the 'release' variant.

PiperOrigin-RevId: 573287429