Add shadow for AlwaysOnHotwordDetector
The goal of this change is to allow testing of a VoiceInteractionService that calls createAlwaysOnHotwordDetector and have the detector return true when startRecognition is called.

I've removed some unneeded shadowing in ShadowVoiceInteractionService in favor of real code. For the detector, there are two critical fields that I'm now setting when the object is created. Note that a test hoping to achieve this result should also provide some fake metadata in its manifest to convince KeyphraseEnrollmentInfo that an enrollment apk is present.

PiperOrigin-RevId: 574140904