Fix zero-width text views for LEGACY graphics in Android U+
Previously, in Android U, for LEGACY graphics, TextViews were always considered
to be zero-width. Add a shadow implementation for Paint.nGetRunCharacterAdvance
in U. This logic is used in the TextView width calculation. Also preemptively
add support for it in V, where the signature has already changed.

This also fixes the Robolectric EspressoTest when running with LEGACY graphics.

PiperOrigin-RevId: 576275359