Introducing SensorEventBuilder to replace ShadowSensorManager.createSensorEvent()
This change refactors logic to create SensorEvents into its own class and adds the ability to set a specific Sensor instance on the SensorEvent. Before this change, users could specify only the type, and a new Sensor instance would be created. Any comparisons on the generated event.sensor property would fail deep equality checks because each event was given a new Sensor instance. It was as if the newly created SensorEvent came from a new Sensor source each time. Setting the whole Sensor allows users to create a SensorEvent that truly behaves like it came from the desired Sensor.

PiperOrigin-RevId: 577906129