Fix ShadowPixelCopy for projects with compileSdk < 34
In Android U, a new PixelCopy.request method that takes a PixelCopy.Request as
a parameter. However, the PixelCopy.Request class was only added in SDK 34.

This will cause a NoClassDefFoundError if PixelCopy is used and the compileSdk is less than 34. Update ShadowPixelCopy to use looseSignatures for this method.

Fixes #8577

PiperOrigin-RevId: 577907854