Fix AttributeSetBuilder lookups with sparse resource table encoding
There were two issues when using AttributeSetBuilder with sparse resource table
encoding. The first was that the offset to the sparse table index was not
taking the header size into account. The second issue was that when a matching
sparse entry was located, it was returning the entry number, not the entry
index, which requires a lookup in the sparse entry index.

Fixes #8578

PiperOrigin-RevId: 578082446