Allowing ShadowSensorManager to have multiple Sensors of the same type added
This is possible in Android, as true uniqueness between sensors comes from the pairing of a Sensor's name and type together. Changing how ShadowSensorManager stores the added Sensor instances will allow for multiple of the same time type to exist at once, more closely acting like the real thing.

Also adding an overload to sendSensorEventToListeners() method to specify which listeners to push a given SensorEvent to. If you create and register multiple Sensors with ShadowSensorManager then want to push SensorEvents to them, it makes sense to only push it to the listeners that are relevant to the given Sensor

PiperOrigin-RevId: 580322304