Add support for setting the behavior of mustPlayShutterSound
MediaActionSound contains a static method mustPlayShutterSound which determines if the hardware camera requires a shutter sound during capture. Behavior of this method relies on the device's underlying AudioManager and AudioService which make it difficult to use in tests. This change adds the ability to override this method so that tests can set the return value.

PiperOrigin-RevId: 585146692