Add support for isLeAudioSupported()
This CL adds support for isLeAudioSupported() method in BluetoothAdapter. If the original bluetooth adapter is used in unit tests, then to verify if Le Audio is supported or not, it depends on the bluetooth service and tests fail as there is no control on bluetooth service in test files. Adding setter and getter functions for Le Audio in ShadowBluetoothAdapter helps to overcome this issue.

PiperOrigin-RevId: 586185847