Standardize on fully-qualified class names for shadow picker map keys
There are two code paths in the Robolectric Processor that adds shadow pickers.
One is the typical path where the shadowed type is known to be present in the
latest SDK level. In this path, the shadow picker map in the generated shadow
provider uses the fully-qualified shadowed class name as the key.

In the other path, when `maxSdk` is used, the dot syntax was being used for the
class name key in the shadow picker map. Update it to use fully qualified names
to be consistent with the case when the shadowed class can be resolved.

This only comes up when shadowing inner classes where the dot syntax is
different than the FQ syntax.

PiperOrigin-RevId: 587946320