Ensure that RNG is loaded when SystemFonts.getAvailableFonts is called in Q/R
In Q and R, calling SystemFonts.getAvailableFonts does not automatically result
in RNG getting loaded like in S+. This is because there are no native calls
before the font directory environment variable is checked. Update the
SystemFonts.buildSystemFallback shadow method to ensure that RNG is loaded
before checking the extracted font directory.

PiperOrigin-RevId: 592757670