Update native method instrumentation to include real native variants
Previously, when a native method was instrumented in Robolectric, it was split
into two methods. For instance, the native method Matrix.nIsIdentity was split
into:

1) The invokedynamic delegator, which contained the original method name:
```
  private static boolean nIsIdentity(long);
    Code:
       0: lload_0
       1: invokedynamic #776,  0            // InvokeDynamic #63:nIsIdentity:(J)Z
       6: ireturn
```

2)The no-op variant, with 'native' removed and zero-value returned
```
  private static final boolean $$robo$$android_graphics_Matrix$nIsIdentity(long);
    Code:
       0: iconst_0
       1: ireturn
```

By default the no-op variants of native methods are invoked to avoid
UnsatisfiedLinkErrors.

This CL introduces a third variant, the 'native' variant, that preserves the
'native' keyword:

3. The native binding method
```
  private static native final boolean $$robo$$nIsIdentity$nativeBinding(long);
```

The goal of this is that, starting in Android V, a set of native libraries will
be included in android-all Jars that contains the native implementations of
SQLite, graphics, and potentially other native functionality, such as
resources.  During JNI registration, these native libraries will bind to these
native binding methods. This will allow Robolectric to selectively invoke
either the no-op variants, the native variants, or any shadows (such as custom
shadows) that want to shadow the native methods.

By default the no-op variants are still used. The native variants will only be
enabled for SQLite and graphics classes starting in Android V.

Also disable preinstrumented jars in Robolectric's GitHub CI until we
can regenerate them.

PiperOrigin-RevId: 592878238