ShadowTelephonyManager instances remember their ids
Methods that used to ignore subId, phoneId, slotIndex, or PhoneAccountHandle parameters will now redirect to the proper TelephonyManager instance, making a new instance if needed.

Methods that used hardcoded 0 as the id now use the associated id if any, but still fall back to 0 to preserve legacy tests.

#getCurrentPhoneType now reads from the preexisting map. This was a preexisting bug revealed by this CL.

The shadowForSubId(...) and similar methods are public to make it trivial for users to correctly create new instances.

Added a #context() method to make it easier/faster to access the context for each TelephonyManager instance.

PiperOrigin-RevId: 596083118