Bump preinstrumented version to 5
Enough changes have accumulated in Robolectric's instrumentation that a version
bump is required. This is mostly due to the updated native method
instrumentation.

Version 5 of the preinstrumented jars have already been deployed to maven.

Also, include a workaround to the preinstrumented build.gradle that fixes the
ability for preinstrumented jars to be signed.

PiperOrigin-RevId: 597403034