Fix double-free in ShadowNativeAllocationRegistryTest.applyFreeFunction_matrix
The first free occurred when applyFreeFunction was explicitly called on the
Matrix native pointer. The second occurred when applyFreeFunction was called
during GC. I incorrectly assumed that zero'ing the pointer would prevent that.

This test segfaulted when run on GitHub CI with an unclear cause.

Also bump the nativeruntime-dist-compat version to 1.0.7 to fix
ShadowNativeBitmapTest.testSameAs_hardware.

PiperOrigin-RevId: 597583110