Support getRoutedDevice and associated listeners for ShadowAudioTrack
We need to handle the listeners in the shadow class because the real
AudioTrack relies on native code to change mutable fields of a private
inner class.

Tests also typically don't have access to the AudioTrack instance, so
the routing is configured statically.

The AudioDeviceInfoBuilder is extended to set some fields that are
required to be non-null, which is now relevant because the new code
calls AudioDeviceInfo.equals.

PiperOrigin-RevId: 599466173