Add a no-op implementation for ShadowNativeBitmap.nativeExtractGainmap
When Bitmap.getGainmap is called, there is a call to
Bitmap.nativeExtractGainmap native code to check if a Gainmap exists.

In a forthcoming change, calls to native methods such as nativeExtractGainmap
will invoke native code directly. A no-op implementation must be provided to
avoid an UnsatisfiedLinkError in U and below.

PiperOrigin-RevId: 599746970