Invoke the embedded Robolectric Native Runtime starting in Android V
Starting in Android V, the native libraries for the Robolectric Native Runtime
will be built from the corresponding framework sources and included in the
android-all Jar. This is to maximize the fidelity of native code for that
particular Android version. Update the native-backed shadows to add the
'callNativeMethodsByDefault' annotation parameter and stop shadowing most native
methods starting in Android V.

Previously, the native code from Android S was being used to simulate native
method calls above Android S. However, as new native functionality was being
added, this was becoming increasingly difficult.

PiperOrigin-RevId: 601167078