Make ShadowDisplayManager report causes of double-initialization
Before this, if someone made a `Shadow#reset` method that called `RuntimeEnvironment.getApplication()`, then `ShadowDisplayManager` would crash with the exception `this method should only be called by Robolectric` and an internal callstack, which was confusing.  With this change, it now reports a "Caused By" exception that shows the `Shadow#reset` callstack, so the developer can see where the error occured.

PiperOrigin-RevId: 601191501