Remove exemptionlist-based throw-on-native method functionality
Now that native-backed shadows for graphics and SQLite are using
callNativeMethodsByDefault, any native methods that are not properly shadowed
will automatically throw an UnsatisfiedLinkError,

This makes the exemptionlist-based system for throwing on native methods
redundant as an UnsatisfiedLinkError will be thrown instead of a
NativeMethodNotFoundException.

The only potential drawback is that throw-on-native functionality will not be
enabled for legacy graphics and legacy SQLite shadows. However:

1) Legacy graphics and SQLite shadows are in maintenance mode and no new
functionality is being added to them.

2) The exemptionlist-based system could not differentiate between legacy and
native-backed shadows, which meant that exemptions were often added for legacy
shadows, and prevented throw-on-native from occurring in native-backed shadows,
where it was most useful.

3) It would be straightforward to add a warning for legacy shadows of native
methods at runtime. This is because ShadowWrangler now has information about
whether a method was originally native or not.

PiperOrigin-RevId: 601474699