Avoid attempting to unpause a paused background Looper that is already quit
There was race condition that could occur in the ShadowPausedLooper resetter.
If a paused background Looper was quit but the backing thread was still alive,
the resetter would attempt to unpause it, but the unpause control message would
never run because the looper was quit. This caused a deadlock while waiting for
the unpause control message to run.

PiperOrigin-RevId: 603412605