Lazily attach a system resource config change listener
Previously, RoboMonitoringInstrumentation.onCreate invoked Resources.getSystem,
which resulted in an AssetManager object being created that encapsulated system
assets very early in the test process. While this is generally innocuous,
because AssetManager caching is currently broken, it caused unnecessary
overhead for tests that do not require interacting with resources.

There will be forthcoming CLs to fix AssetManager caching and filtering, but
avoiding referencing system resources so early in a test method cycle a good
for tests that do not need resources.

PiperOrigin-RevId: 605909387