Fix system AssetManager pointer cache
Previously, both ShadowArscAssetManager9 and ShadowArscAssetManager10 contained
identical caching mechanisms for system AssetManager pointers. However, this
caching mechanism was convoluted and incorrect. It assumed that the first
invocation to the AssetManager(boolean) constructor was for non-system assets
(presumably Application assets).  However, this assumption does not hold true
anymore, beacuse the first invocation is for system assets.

For tests using lazy Application, the AssetManager pointer for system assets
was always being recreated.

Update the caching mechanism to hook into the call to AssetManager.getSystem to
determine when system assets are being created. This is only called when the
System assetManager is requested. Add a test for this as well.

PiperOrigin-RevId: 606282824