Improve the performance of RebuildFilterList by 6x
In the main loop of CppAssetManager2.RebuildFilterList, each ResTable_config
object was cloned before the match function, which was unnecessary if the
ResTable_config was not a match. During the filter process, the majority of
entries are not matches (only around 100 out of ~4000 are matches).

This caused allocation of ~4000 ResTable_config objects each time, even though
only a small fraction matched.

Update the logic to only make a copy if the match is successful.

This makes RebuildFilterList around 6x faster. In the added benchmark, the
time of the benchmark went from around 6s -> 1s.

PiperOrigin-RevId: 606294716