Avoid system resource configuration update thrashing
Previously, at the beginning of every test, when Resources.getSystem was
called, the system Resources was initialized with an empty Configuration and
then immediately updated with the 'Bootstrap' configuration. This ended up
rebuilding the system asset manager filter lists at least twice during each
test, which was costly for test classes with a large number of test methods.
Update the system Resources initialization logic to be set to the 'Bootstrap'
configuration upfront to avoid the thrashing.

PiperOrigin-RevId: 606643334