Fix ShadowBluetoothGatt.setCharacteristicNotification implementation
The implementation currently does not match the documentation. Specifically unsubscribing from a characteristic is only allowed if disallowCharacteristicNotification was called.

The intention here seems to be that any call to setCharacteristicNotification should be allowed if allowCharacteristicNotification was called, and all calls should be blocked in disallowCharacteristicNotification was called. Update the implementation to match this.

PiperOrigin-RevId: 606870125