Added code to stop gl thread
 terminate the thread in callOnDetachFromWindow

Added code to stop gl thread

I added fields to hold the renderer and the GL thread.

In onPause(), I interrupt the GL thread if it's not null.

In onResume(), I start a new GL thread only if the renderer is not null.

I added a setRenderer() method to set the renderer externally.

I introduced a reset() method, annotated with @Resetter, which you can use to clean up resources and terminate the GL thread if needed after each test.

 terminate the thread in callOnDetachFromWindow

Update ShadowGLSurfaceView.java