Clear Choreographer.getInstance for all Loopers.
Previously Robolectric would attempt to clear Choreographer.getInstance in a resetter method.
However, the Choreographer instance is a ThreadLocal, so when the resetter would have no effect when
running in INSTRUMENTATION_TEST LooperMode, or if a Choreographer was used on another Looper thread.
This can lead to hard-to-debug test pollution issues where Choreographer vsync callbacks don't occur as expected.

This commit changes the reset logic to clear Choreographer.getInstance for all active Looper threads.

PiperOrigin-RevId: 610885585