Use real Android View scrolling methods by default
In Robolectric, when using legacy graphics mode, many scrolling-related methods
were using an arbitrary fake implementation that was added over ten years ago.
This deviated in several ways from real Android, such as:

* Allowed setting scroll positions on views with no children.

* Allowed setting scroll positions on views that are not scrollable, i.e. the
child content does not exceed the bounds of the scrollable parent.

* Allowed setting scroll positions on views detached from Activity hierarchies.

In order to make scrolling-related testing as high fidelity as possible, update
the default scroll behavior to use the real Android code.

For tests that depend on the legacy/broken scrolling logic, it is simple to use
the old behavior by using the 'robolectric.useRealScrolling=false' system
property. It is highly recommended not to use this property though in order to
have tests be more robust and high-quality.

PiperOrigin-RevId: 612519883