Update ShadowBackupManager to support testing non-successful restores
Restore sets in the ShadowBackupManager now also contain the result that should
be passed to the restoreFinished callback at the end of a restore. This makes
it possible to test restore failures.

In addition, to allow the behaviour to change between subsequent restore calls,
it's also possible to provide multiple restore sets for the same restore token.
If multiple restore sets were added for a token, every call to
`#restorePackages` or `#restorePackage` removes the set it operates on except
if it was the last one.

The ShadowBackupManager now also keeps track of the number of times a package
was restored.

PiperOrigin-RevId: 614587898