Make it possible for `ShadowBackupManager` to return `null` restore sets.
It's possible for the `RestoreObserver#restoreSetsAvailable` callback to be
passed `null` instead of an empty array. Some backup transports may do this to
distinguish between successfully fetching an empty restore set and failing to
fetch a restore set.

This change makes it possible to tell the `ShadowBackupManager` that a `null`
restore set should be returned in the callback as a result of calling
`getAvailableRestoreSets`.

PiperOrigin-RevId: 617849938