Add logic to remove obsolete temp directories in Windows
In Windows, unlike Linux and Mac, it is generally not possible to remove open
files that are in use by the system.

When the Robolectric Native Runtime (RNR) is loaded in Windows, there are two
files that cannot be deleted while the current JVM process is running. The
first is robolectric-nativeruntime.dll, which is loaded using System.load, and
the second is the ICU dat file, which is mapped to memory. On Linux and Mac, it
is possible to delete these files in the current running JVM process. However,
on Windows, trying to delete these will not succeed.

As a workaround, add logic to mark these obsolete temp file directories. If
Robolectric is started again, add some logic to detect unused nativeruntime
temp directories and clean them up in the TempDirectory shutdown hook.

PiperOrigin-RevId: 617877469