Support KeyEvent#keyCodeFromString.
Starting in Android API 21, the implementation of keyCodeFromString moved from
java to native. This resulted in keyCodeFromString typically always
returning an incorrect value of 0 when run in Robolectric for SDKs >= 21.

This commit adds support by building a map of keycode strings to ints by reflectively
inspecting all KeyEvent.KEYCODE_ fields. And adds a ctesque test to verify correctness.

PiperOrigin-RevId: 619240067